requirements.txt
inference
supervision
opencv-python
huggingface_hub
ultralytics
pillow
roboflow